Changed package structure to follow standared maven which will be added in the next pull request soon
