+++ modules/cloud/AuroraIntegration/src/main/java/org/apache/airavata/cloud/exceptions/auroraExceptions/AuroraException.java
@@ -1,0 +1,15 @@
+package main.java.org.apache.airavata.cloud.exceptions.auroraExceptions;
+
+public class AuroraException extends Exception {
+    private String exceptionMsg;
+    public AuroraException(){
+    	exceptionMsg="";
+    }
+    public AuroraException(String exceptionMsgIn){
+    	exceptionMsg=exceptionMsgIn;;
+    }
+
+    public String toString(){
+    	return 	this.exceptionMsg;
+    }
+ }
