+++ modules/cloud/AuroraIntegration/src/main/java/org/apache/airavata/cloud/utilities/auroraUtilities/AuroraUtilI.java
@@ -1,0 +1,10 @@
+package main.java.org.apache.airavata.cloud.utilities.auroraUtilities;
+
+import java.io.BufferedReader;
+
+import main.java.org.apache.airavata.cloud.exceptions.auroraExceptions.AuroraException;
+
+public interface AuroraUtilI{
+  public void printLog(BufferedReader stdout) throws AuroraException;
+  public BufferedReader executeProcess(String commandToRunProcess) throws AuroraException;
+}
