+++ modules/cloud/AuroraIntegration/src/main/java/org/apache/airavata/cloud/utilities/marathonUtilities/MarathonUtilI.java
@@ -1,0 +1,11 @@
+package main.java.org.apache.airavata.cloud.utilities.marathonUtilities;
+
+import java.io.BufferedReader;
+
+import main.java.org.apache.airavata.cloud.exceptions.auroraExceptions.AuroraException;
+
+public interface MarathonUtilI{
+
+  //TODO:Abstract Marathon methods goes heresss
+
+}
