+++ modules/cloud/AuroraIntegration/src/main/java/org/apache/airavata/cloud/utilities/marathonUtilities/MarathonUtilImpl.java
@@ -1,0 +1,14 @@
+
+package main.java.org.apache.airavata.cloud.utilities.marathonUtilities;
+
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+
+import main.java.org.apache.airavata.cloud.exceptions.marathonExceptions.MarathonException;
+
+public class MarathonUtilImpl implements MarathonUtilI{
+  //TODO: Marathon utilities code goes here
+
+}
