--- modules/cloud/AuroraIntegration/src/org/apache/airavata/cloud/exceptions/auroraExceptions/AuroraException.java
@@ -1,15 +1,0 @@
-package org.apache.airavata.cloud.exceptions.auroraExceptions;
-
-public class AuroraException extends Exception {
-    private String exceptionMsg;
-    public AuroraException(){
-    	exceptionMsg="";
-    }
-    public AuroraException(String exceptionMsgIn){
-    	exceptionMsg=exceptionMsgIn;;
-    }
-
-    public String toString(){
-    	return 	this.exceptionMsg;
-    }
- }
