--- modules/cloud/AuroraIntegration/src/org/apache/airavata/cloud/utilities/marathonUtilities/MarathonUtilI.java
@@ -1,11 +1,0 @@
-package org.apache.airavata.cloud.utilities.marathonUtilities;
-
-import java.io.BufferedReader;
-
-import org.apache.airavata.cloud.exceptions.auroraExceptions.AuroraException;
-
-public interface MarathonUtilI{
-
-  //TODO:Abstract Marathon methods goes heresss 
-
-}
