--- modules/cloud/AuroraIntegration/src/org/apache/airavata/cloud/utilities/marathonUtilities/MarathonUtilImpl.java
@@ -1,14 +1,0 @@
-
-package org.apache.airavata.cloud.utilities.marathonUtilities;
-
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-
-import org.apache.airavata.cloud.exceptions.marathonExceptions.MarathonException;
-
-public class MarathonUtilImpl implements MarathonUtilI{
-  //TODO: Marathon utilities code goes here
-
-}
